move infomation into scripted agent
