+++ nidup/pysc2/agent/scripted/information.py
